\name{EasyFlow2}
\alias{EasyFlow2}
\alias{EasyFlow2-package}
\alias{Annotate_Clusters_Dual_Track}
\alias{Apply_Gating_Template}
\alias{Apply_Panel_Mapping}
\alias{Apply_Plot_Order}
\alias{Assign_Terminal_Gate_CellType}
\alias{Auto_Clean_FCS}
\alias{Build_Cluster_Summary_Values}
\alias{Call_OpenRouter_Chat}
\alias{Compare_Populations}
\alias{Compare_Populations_Barplot}
\alias{Export_To_FCS}
\alias{Generate_Final_HTML_Report}
\alias{Infer_Trajectory}
\alias{Match_Rules_With_Negatives}
\alias{Normalize_Annotation_Label}
\alias{Perform_Abundance_Stats}
\alias{Perform_Marker_DE}
\alias{Plot_Alluvial}
\alias{Plot_Composition}
\alias{Plot_Confusion_Matrix}
\alias{Plot_Density_Line}
\alias{Plot_Dim_Reduction}
\alias{Plot_Dotplot}
\alias{Plot_Gating_Tree}
\alias{Plot_Heatmap}
\alias{Plot_Sample_PCA}
\alias{Plot_Scatterhist_Line}
\alias{Plot_Violin_AllInOne}
\alias{Plot_tSNE_Density}
\alias{Prepare_Raw_Data}
\alias{Project_Atlas_To_Massive_Data}
\alias{Project_New_Sample}
\alias{Remove_Batch_Effects_ADTnorm_Lite}
\alias{Run_Agentic_LLM_Loop}
\alias{Run_Analysis_On_Gated_Data}
\alias{Run_Universal_Gating}
\alias{Set_Plot_Order}
\alias{theme_publication}
\docType{package}
\title{EasyFlow2: Streamlined Flow Cytometry Analysis Workflow}
\description{
EasyFlow2 provides an end-to-end flow cytometry workflow for FCS preprocessing,
interactive gating, dimensionality reduction, clustering, annotation, atlas
projection, visualization, statistics, trajectory inference, FCS export, and
HTML reporting.
}
\details{
See the repository README and the integrated workflow files installed under
\code{system.file("workflows", package = "EasyFlow2")} for full usage examples.
}
\keyword{package}
