% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Plot_Heatmap}
\alias{Plot_Heatmap}
\title{绘制全能热图 (支持自定义参数)}
\usage{
Plot_Heatmap(
  res_obj,
  group_by = "Cluster",
  scale = "row",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  color_palette = NULL,
  show_border = TRUE,
  fontsize = 10
)
}
\description{
绘制全能热图 (支持自定义参数)
}
