% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Prepare_Raw_Data}
\alias{Prepare_Raw_Data}
\title{快速读取并合并 FCS 文件}
\usage{
Prepare_Raw_Data(work_dir, group_keywords, n_cells_per_sample = 5000)
}
\arguments{
\item{work_dir}{FCS 文件夹路径}

\item{group_keywords}{分组关键词 (如 c("Ctrl", "Treat"))}

\item{n_cells_per_sample}{每个样本抽样数 (建议 5000-50000)}
}
\description{
快速读取并合并 FCS 文件
}
