% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Run_Analysis_On_Gated_Data}
\alias{Run_Analysis_On_Gated_Data}
\title{自动分析圈好的数据}
\usage{
Run_Analysis_On_Gated_Data(gated_data, k_neighbors = 30)
}
\arguments{
\item{gated_data}{圈门后的数据}

\item{k_neighbors}{聚类精度}
}
\description{
自动分析圈好的数据
}
