% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Run_Automated_Analysis}
\alias{Run_Automated_Analysis}
\title{启动全自动分析流程}
\usage{
Run_Automated_Analysis(
  work_dir,
  group_keywords,
  n_cells_per_sample = 3000,
  k_neighbors = 30
)
}
\arguments{
\item{work_dir}{FCS 文件夹路径}

\item{group_keywords}{分组关键词}

\item{n_cells_per_sample}{每个样本抽样数}

\item{k_neighbors}{聚类精度}
}
\description{
启动全自动分析流程
}
